pyproject.toml
[tool.poetry]
name = "studio"
version = "0.4.0"
description = "Demo application to showcase the Arcane Development Kit"
authors = ["Marco lamina <marco@arcane.engineer>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.10.0"
django = "^5.1.1"
arcane-engine = "^1.10.4"
pygithub = "^2.4.0"
markdown = "^3.7"
django-allauth = "^65.0.1"
psycopg2-binary = "^2.9.9"
django-redis = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
